move save simulator tests to other file
